Applied book-qut theme changes to book and removed book-qut
