--- themes/book-qut/.eslintrc.js
@@ -1,4 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-  extends: ["@remix-run/eslint-config", "@remix-run/eslint-config/node"],
-};
