--- themes/book-qut/.gitignore
@@ -1,21 +1,0 @@
-node_modules
-
-.cache
-.env
-.vercel
-.output
-.turbo
-.yalc
-yalc.lock
-
-/public/build
-/public/*thebe*
-/public/service-worker-*.js
-/public/service-worker.js
-/api/*
-/build/*
-
-/app/styles/app.css
-
-# This is ignored as it is built
-remix.config.js
