--- themes/book-qut/README.md
@@ -1,28 +1,0 @@
-# @myst-theme/book
-
-A lightweight MyST theme designed to mimic the look-and-feel of an interactive book.
-
-## Usage with MyST
-
-To use this template locally, update your site template make sure your project's local dependencies are installed:
-
-```sh
-site:
-  template: book-theme
-```
-
-Then start the local server:
-
-```sh
-myst start
-```
-
-Open up [http://localhost:3000](http://localhost:3000) and you should be ready to go!
-
-## Development
-
-This repository depends on the packages in [myst-theme](https://github.com/executablebooks/myst-theme) and is best developed in that repository to allow for live-reload, etc.
-
-## Deployment
-
-See the [Remix docs](https://remix.run/docs) for more information on custom deployments.
