--- themes/book-qut/app/root.tsx
@@ -1,104 +1,0 @@
-import type { LinksFunction, V2_MetaFunction, LoaderFunction } from '@remix-run/node';
-import tailwind from '~/styles/app.css';
-import thebeCoreCss from 'thebe-core/dist/lib/thebe-core.css';
-import { getConfig } from '~/utils/loaders.server';
-import type { SiteLoader } from '@myst-theme/common';
-import {
-  Document,
-  responseNoSite,
-  getMetaTagsForSite,
-  getThemeSession,
-  ContentReload,
-  SkipTo,
-} from '@myst-theme/site';
-export { AppErrorBoundary as ErrorBoundary } from '@myst-theme/site';
-import { createSearch as createMiniSearch } from '@myst-theme/search-minisearch';
-import { Outlet, useLoaderData } from '@remix-run/react';
-import { SearchFactoryProvider } from '@myst-theme/providers';
-import type { ISearch, MystSearchIndex } from '@myst-theme/search';
-import { SEARCH_ATTRIBUTES_ORDERED } from '@myst-theme/search';
-import { useCallback, useMemo } from 'react';
-
-export const meta: V2_MetaFunction<typeof loader> = ({ data }) => {
-  return getMetaTagsForSite({
-    title: data?.config?.title,
-    description: data?.config?.description,
-    twitter: data?.config?.options?.twitter,
-  });
-};
-
-export const links: LinksFunction = () => {
-  return [
-    {
-      rel: 'icon',
-      href: '/favicon.ico',
-    },
-    { rel: 'stylesheet', href: tailwind },
-    { rel: 'stylesheet', href: thebeCoreCss },
-    { rel: 'stylesheet', href: '/myst-theme.css' },
-    {
-      rel: 'stylesheet',
-      href: 'https://cdn.jsdelivr.net/npm/jupyter-matplotlib@0.11.3/css/mpl_widget.css',
-    },
-    {
-      rel: 'stylesheet',
-      href: 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css',
-    },
-  ];
-};
-
-export const loader: LoaderFunction = async ({ request }): Promise<SiteLoader> => {
-  const baseURL = process.env.BASE_URL || undefined;
-  const [config, themeSession] = await Promise.all([
-    getConfig().catch(() => null),
-    getThemeSession(request),
-  ]);
-  if (!config) throw responseNoSite();
-  const data = {
-    theme: themeSession.getTheme(),
-    config,
-    CONTENT_CDN_PORT: process.env.CONTENT_CDN_PORT ?? 3100,
-    MODE: (process.env.MODE ?? 'app') as 'app' | 'static',
-    BASE_URL: baseURL,
-  };
-  return data;
-};
-
-function createSearch(index: MystSearchIndex): ISearch {
-  const options = {
-    fields: SEARCH_ATTRIBUTES_ORDERED as any as string[],
-    storeFields: ['hierarchy', 'content', 'url', 'type', 'id', 'position'],
-    idField: 'id',
-    searchOptions: {
-      fuzzy: 0.2,
-      prefix: true,
-    },
-  };
-  return createMiniSearch(index.records, options);
-}
-
-export default function AppWithReload() {
-  const { theme, config, CONTENT_CDN_PORT, MODE, BASE_URL } = useLoaderData<SiteLoader>();
-
-  const searchFactory = useCallback((index: MystSearchIndex) => createSearch(index), []);
-
-  return (
-    <SearchFactoryProvider factory={searchFactory}>
-      <Document
-        theme={theme}
-        config={config}
-        scripts={MODE === 'static' ? undefined : <ContentReload port={CONTENT_CDN_PORT} />}
-        staticBuild={MODE === 'static'}
-        baseurl={BASE_URL}
-      >
-        <SkipTo
-          targets={[
-            { id: 'skip-to-frontmatter', title: 'Skip to article frontmatter' },
-            { id: 'skip-to-article', title: 'Skip to article content' },
-          ]}
-        />
-        <Outlet />
-      </Document>
-    </SearchFactoryProvider>
-  );
-}
