--- themes/book-qut/app/routes/[sitemap_style.xsl].tsx
@@ -1,1 +1,0 @@
-export { sitemapStylesheetResponse as loader } from '@myst-theme/site';
