--- themes/book-qut/app/routes/api.theme.tsx
@@ -1,1 +1,0 @@
-export { setThemeAPI as action } from '@myst-theme/site';
