--- themes/book-qut/app/types.ts
@@ -1,8 +1,0 @@
-export interface TemplateOptions {
-  hide_toc?: boolean;
-  hide_outline?: boolean;
-  hide_search?: boolean;
-  hide_footer_links?: boolean;
-  outline_maxdepth?: number;
-  hide_title_block?: boolean;
-}
