--- themes/book-qut/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@eresearchqut/myst-theme-book",
-  "version": "0.13.3",
-  "private": true,
-  "sideEffects": false,
-  "scripts": {
-    "compile": "tsc --noEmit --skipLibCheck",
-    "build:css": "tailwindcss -m -i ./styles/app.css -o app/styles/app.css",
-    "dev:copy": "cp remix.config.dev.js remix.config.js",
-    "prod:copy": "cp remix.config.prod.js remix.config.js",
-    "clean": "rimraf public/build build api",
-    "build:thebe": "node ../../bin/copy-thebe-assets.cjs ./public",
-    "prod:build": "npm run prod:copy && npm run build:thebe && npm run build:css && remix build",
-    "dev:css": "tailwindcss -w -i ./styles/app.css -o app/styles/app.css",
-    "dev": "npm run dev:copy && npm run build:thebe && concurrently \"npm run dev:css\" \"remix dev\"",
-    "start": "npm run build:css && remix dev"
-  },
-  "dependencies": {
-    "@myst-theme/icons": "^0.13.3",
-    "@myst-theme/jupyter": "^0.13.3",
-    "@myst-theme/providers": "^0.13.3",
-    "@myst-theme/site": "^0.13.3",
-    "@myst-theme/common": "^0.13.3",
-    "@myst-theme/styles": "^0.13.3",
-    "@remix-run/node": "~1.17.0",
-    "@remix-run/react": "~1.17.0",
-    "@remix-run/vercel": "~1.17.0",
-    "myst-common": "^1.7.3",
-    "myst-config": "^1.7.3",
-    "node-fetch": "^2.6.11",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@remix-run/dev": "~1.17.0",
-    "@remix-run/eslint-config": "~1.17.0",
-    "@remix-run/serve": "~1.17.0",
-    "@tailwindcss/typography": "^0.5.9",
-    "@types/node-fetch": "^2.6.4",
-    "@types/react": "^18.2.24",
-    "@types/react-dom": "^18.2.8",
-    "@vercel/node": "^2.15.1"
-  },
-  "engines": {
-    "node": ">=14"
-  },
-  "packageManager": "npm@8.10.0"
-}
