--- themes/book-qut/remix.config.dev.js
@@ -1,72 +1,0 @@
-/** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
-  publicPath: '/myst_assets_folder/',
-  serverBuildPath: 'api/index.js',
-  serverMainFields: ['main', 'module'],
-  serverModuleFormat: 'cjs',
-  serverPlatform: 'node',
-  serverMinify: false,
-  // When running locally in development mode, we use the built in remix
-  // server. This does not understand the vercel lambda module format,
-  // so we default back to the standard build output.
-  server: process.env.NODE_ENV === 'development' ? undefined : './server.js',
-  ignoredRouteFiles: ['**/.*'],
-  serverDependenciesToBundle: [
-    /^rehype.*/,
-    /^remark.*/,
-    /^unified.*/,
-    /^unist.*/,
-    /^hast.*/,
-    /^mdast.*/,
-    /^micromark.*/,
-    /^csv-parse.*/,
-    'html-whitespace-sensitive-tag-names',
-    'html-void-elements',
-    'property-information',
-    'array-iterate',
-    'stringify-entities',
-    'comma-separated-tokens',
-    'trim-trailing-lines',
-    'escape-string-regexp',
-    'ccount',
-    'web-namespaces',
-    'space-separated-tokens',
-    'character-entities-legacy',
-    'character-entities-html4',
-    'trim-lines',
-    'bail',
-    'is-plain-obj',
-    'trough',
-    'zwitch',
-    'nanoid',
-    /^vfile.*/,
-    /^myst-.*/,
-    'markdown-it-myst',
-    'simple-validators',
-    'doi-utils',
-    'orcid',
-    'credit-roles',
-    'tex-to-typst',
-    'jats-tags',
-    /^@myst-theme\/.*/,
-    /react-syntax-highlighter.*/,
-    'markdown-it-myst-extras',
-    'markdown-it-dollarmath',
-    'markdown-it-amsmath',
-    'swr',
-    'swr/immutable',
-    'devlop',
-    '@curvenote/ansi-to-react',
-    'jats-utils',
-    'react-merge-refs',
-  ],
-  watchPaths: ['../../packages/**/*'],
-  future: {
-    v2_dev: true,
-    v2_routeConvention: true,
-    v2_normalizeFormMethod: true,
-    v2_headers: true,
-    v2_meta: true,
-    v2_errorBoundary: true,
-  },
-};
