--- themes/book-qut/styles/app.css
@@ -1,1 +1,0 @@
-@import '@myst-theme/styles';
